Initial MVP
Prompt:
You are an lead engineer with expert knowledge on React and NodeJS. You have give a task to create a webapp that allows users to create a cv/resume in markdown format and see it rendered in to a beautiful cv/resume. Users should also able to download a pdf version of the resume as well. The generated pdf should be ATS friendly. You will be using NextJS and React for this web app you are free to choose other required dependencies to provide the above functionalities.

Before starting any development we need to comeup with a plan for this project. Lets start by creating that
---
we need a feature to allow adding a professinal headshot of the user to the cv and also user should have a option to select a 2 column layout as well where first column with a more witdh containing key sections like summary, experience and second right colun containing info like headshot, contact details, skills, education
---
rather than allowing image upload what do you think about proving the image using a markdown link?
---
I like this plan. Let's first document this so we can keep ourselves aligned to this while we work
---
yes let's begin with phase 1 of our plan